Refactor migrations.

+++ database/migrations/2016_10_31_160156_create_invoices_table.php
@@ -16,7 +16,7 @@ class CreateInvoicesTable extends Migration
         Schema::create('invoices', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title');
-            $table->decimal('price', 2);
+            $table->double('price', 2);
             $table->boolean('paid');
             $table->timestamps();
         });
